Upgrade to rustc 1.0.0-dev (47551b574 2015-04-14) (built 2015-03-18)

+++ src/platform/android/surface.rs
@@ -134,10 +134,10 @@ impl EGLImageNativeSurface {
         }
     }
 
-    pub fn get_id(&self) -> int {
+    pub fn get_id(&self) -> isize {
         match self.image {
             None => 0,
-            Some(image_khr) => image_khr as int,
+            Some(image_khr) => image_khr as isize,
         }
     }
